[WRFE-55](refactor): 로그인/회원가입 폴더구조 리팩토링
