+++ apps/front/wraffle-webview/src/entities/auth/index.ts
@@ -1,1 +1,1 @@
-export * from './login';
+export * from './schema';
