--- apps/front/wraffle-webview/src/entities/auth/join/index.ts
@@ -1,3 +1,0 @@
-export * from './constant';
-export * from './schema';
-export * from './type';
