+++ apps/front/wraffle-webview/src/features/validate-phone/config/const.ts
@@ -1,0 +1,3 @@
+export const DEFAULT_ERROR_MESSAGE = '';
+export const MISMATCHED_CODE_ERROR_MESSAGE = '인증번호가 일치하지 않습니다.';
+export const VERIFY_CODE_LENGTH = 4;
