+++ apps/front/wraffle-webview/src/features/validate-phone/config/index.ts
@@ -1,0 +1,1 @@
+export * from './const';
