+++ apps/front/wraffle-webview/src/shared/hook/useAutoFocus.ts
@@ -23,4 +23,4 @@ const useAutoFocus = (
   return [ref, handleKeyUp];
 };
 
-export default useAutoFocus;
+export {useAutoFocus};
