+++ apps/front/wraffle-webview/src/shared/ui/form/index.ts
@@ -1,2 +1,3 @@
 export * from './RHF/RHFInput';
 export * from './RHF/core/Form';
+export * from './GenericForm';
