+++ apps/front/wraffle-webview/src/widgets/join/config/funnel.ts
