+++ apps/front/wraffle-webview/src/widgets/join/ui/Name.tsx
@@ -1,5 +1,5 @@
+import type {JoinPayload} from '../config';
 import {useFormContext} from 'react-hook-form';
-import type {JoinPayload} from '@/entities/auth/join/schema';
 import {RHFInput} from '@/shared/ui';
 import {Button, Typography} from '@wraffle/ui';
 
@@ -39,7 +39,11 @@ const Name = ({onNext}: NameProps) => {
       />
 
       <div className='fixed inset-x-0 bottom-0 bg-white p-5'>
-        <Button onClick={() => onNext(name, nickname)} disabled={isDisabled()}>
+        <Button
+          type='button'
+          onClick={() => onNext(name, nickname)}
+          disabled={isDisabled()}
+        >
           다음
         </Button>
       </div>
