+++ apps/front/wraffle-webview/src/widgets/login/config/index.ts
