+++ apps/front/wraffle-webview/src/widgets/login/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './EmailForm';
